@@ -22,7 +22,7 @@ class ThemeNotifier with ChangeNotifier {
     dividerColor: Colors.white54,
   );
 
-  ThemeData _themeData;
+  late ThemeData _themeData;
   ThemeData getTheme() => lightTheme;
 
   ThemeNotifier() {
