@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 class Header extends StatelessWidget {
   final String headerName;
   const Header({
-    Key key,
-    this.headerName,
+    Key? key,
+    required this.headerName,
   }) : super(key: key);
 
   @override
@@ -26,7 +26,7 @@ class Header extends StatelessWidget {
 
 class ProfileCard extends StatelessWidget {
   const ProfileCard({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
