@@ -8,8 +8,8 @@ class MenuController extends ChangeNotifier {
   GlobalKey<ScaffoldState> get scaffoldKey => _scaffoldKey;
 
   void controlMenu() {
-    if (!_scaffoldKey.currentState.isDrawerOpen) {
-      _scaffoldKey.currentState.openDrawer();
+    if (!_scaffoldKey.currentState!.isDrawerOpen) {
+      _scaffoldKey.currentState!.openDrawer();
     }
   }
 }
